Fade to credits and fuse box + material fixes
Added a fuse box and a check that makes sure elevator is fixed before opening the doors. When player inserts mechanical parts into the fuse box, the material in one of the lights changes to green. Made the screen fade to black before starting credits scene. Change and fixed some materials.